moved the findExistingPage function to the utils folder
